Add modules task and view
